fix signed shift in blend32 use typefacecache in fonthost_win
git-svn-id: http://skia.googlecode.com/svn/trunk@1224 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/core/SkBlitter_ARGB32.cpp
@@ -45,10 +45,10 @@ static inline int upscale31To32(int value) {
     return value + (value >> 4);
 }
 
-static inline unsigned blend32(unsigned src, unsigned dst, unsigned scale) {
-    SkASSERT(src <= 0xFF);
-    SkASSERT(dst <= 0xFF);
-    SkASSERT(scale <= 32);
+static inline int blend32(int src, int dst, int scale) {
+    SkASSERT((unsigned)src <= 0xFF);
+    SkASSERT((unsigned)dst <= 0xFF);
+    SkASSERT((unsigned)scale <= 32);
     return dst + ((src - dst) * scale >> 5);
 }
